fix: Refactorizar las rutas de importación para los esquemas de validación
