+++ src/validations/financiero/outcomeSchema.ts
