+++ src/validations/gestion-contenido/crearPreguntaSchema.ts
